oh no

+++ Backend/src/main/java/capstone/p2plend/service/RequestService.java
@@ -7,10 +7,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import capstone.p2plend.entity.Account;
+import capstone.p2plend.entity.User;
 import capstone.p2plend.entity.Deal;
 import capstone.p2plend.entity.Request;
-import capstone.p2plend.repo.AccountRepository;
+import capstone.p2plend.repo.UserRepository;
 import capstone.p2plend.repo.DealRepository;
 import capstone.p2plend.repo.RequestRepository;
 
@@ -21,7 +21,7 @@ public class RequestService {
 	RequestRepository requestRepo;
 
 	@Autowired
-	AccountRepository accountRepo;
+	UserRepository accountRepo;
 
 	@Autowired
 	DealRepository dealRepo;
@@ -31,7 +31,7 @@ public class RequestService {
 
 //	@Autowired
 //    private ModelMapper modelMapper;
-	
+
 	@Transactional
 	public List<Request> findAll() {
 		return requestRepo.findAll();
@@ -41,17 +41,16 @@ public class RequestService {
 	public Request getOneById(int id) {
 		return requestRepo.findById(id).get();
 	}
-	
+
 	@Transactional
 	public List<Request> findAllExceptUserRequest(String token) {
 		List<Request> listRq = new ArrayList<>();
 
 		String username = jwtService.getUsernameFromToken(token);
-		Account account = accountRepo.findByUsername(username);
+		User account = accountRepo.findByUsername(username);
 
 		listRq = requestRepo.findAllUserRequestExcept(account.getId());
-		
-		
+
 		return listRq;
 	}
 
@@ -67,29 +66,29 @@ public class RequestService {
 //	    }
 //	    return post;
 //	}
-	
+
 	@Transactional
 	public List<Request> findAllRequestHistoryDone(String token) {
 		List<Request> listRq = new ArrayList<>();
 
 		String username = jwtService.getUsernameFromToken(token);
-		Account account = accountRepo.findByUsername(username);
+		User account = accountRepo.findByUsername(username);
 		listRq = requestRepo.findAllUserHistoryRequestDone(account.getId(), "done");
-		
+
 		return listRq;
 	}
-	
+
 	@Transactional
 	public boolean createRequest(Request request, String token) {
 		try {
 			String username = jwtService.getUsernameFromToken(token);
-			Account account = accountRepo.findByUsername(username);
+			User account = accountRepo.findByUsername(username);
 			Deal deal = new Deal();
 			deal.setStatus("pending");
 
 			deal.setRequest(request);
 
-			request.setFromAccount(account);
+			request.setBorrower(account);
 			request.setDeal(deal);
 
 			requestRepo.save(request);
@@ -104,8 +103,8 @@ public class RequestService {
 		try {
 			Request existRequest = requestRepo.findById(id).get();
 			String username = jwtService.getUsernameFromToken(token);
-			Account account = accountRepo.findByUsername(username);
-			existRequest.setToAccount(account);
+			User account = accountRepo.findByUsername(username);
+			existRequest.setLender(account);
 			Deal deal = existRequest.getDeal();
 			deal.setStatus("transitioning");
 
@@ -118,8 +117,19 @@ public class RequestService {
 	}
 
 	@Transactional
-	public boolean remove(int id) {
+	public boolean remove(int id, String token) {
 		try {
+			Request request = requestRepo.findById(id).get();
+			User user = accountRepo.findByUsername(jwtService.getUsernameFromToken(token));
+			
+			if(request.getBorrower().getId() != user.getId()) {
+				return false;
+			}
+			
+			if(!request.getDeal().getStatus().equals("pending")) {
+				return false;
+			}
+			
 			requestRepo.deleteById(id);
 			return true;
 		} catch (Exception e) {
